pyproject.toml
[project]
name = "keijiban"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-ai-contentsafety>=1.0.0",
    "dotenv>=0.9.9",
    "pyodbc>=5.2.0",
    "streamlit>=1.47.1",
]
